Address review comments, separate indirect passing differentiation tests.
- Address review comments from @rxwei.
- Create separate indirect passing differentiation tests so that original tests
  aren't affected by workarounds.
